class job_idle moved into a separate file
