+++ src/core/parallel/thread_worker.cc
@@ -28,9 +28,6 @@
 namespace dt {
 
 
-//------------------------------------------------------------------------------
-// thread worker
-//------------------------------------------------------------------------------
 
 /**
   * The worker creates its own thread of execution. The thread will be
@@ -122,134 +119,6 @@ size_t ThreadWorker::get_index() const noexcept {
 
 void ThreadWorker::assign_job(ThreadJob* job) noexcept {
   scheduler = job;
-}
-
-
-
-
-//------------------------------------------------------------------------------
-// "worker controller" scheduler
-//------------------------------------------------------------------------------
-
-idle_job::sleep_task::sleep_task(idle_job* ij)
-  : controller(ij), next_scheduler{nullptr} {}
-
-void idle_job::sleep_task::execute() {
-  controller->n_threads_running--;
-  semaphore.wait();
-  xassert(next_scheduler);
-  thpool->assign_job_to_current_thread(next_scheduler);
-}
-
-
-idle_job::idle_job() {
-  curr_sleep_task = new sleep_task(this);
-  prev_sleep_task = new sleep_task(this);
-  n_threads_running = 0;
-}
-
-
-ThreadTask* idle_job::get_next_task(size_t) {
-  return curr_sleep_task;
-}
-
-
-/**
- * When this method is run, all other threads are inside the
- * `curr_sleep_task->execute()` method. When we modify variables
- * `curr_sleep_task` / `prev_sleep_task`, we do so under the
- * protection of a mutex, ensuring that the other threads cannot
- * attempt to read those variables at this time. The other methods
- * of this class can treat these variables as if they were constant,
- * meaning that they can be safely read even without mutex protection.
- *
- * The second part of this method (after the lock is unlocked) is
- * already multi-threaded: at that point other threads wake up and
- * may call arbitrary API of `idle_job`.
- *
- * Note that we set the variable `n_threads_running` explicitly here
- * (as opposed to, say, allowing each thread to increment this counter
- * upon awaking). This is necessary, because we want to prevent the
- * situation where the OS would delay waking up the threads, so that
- * by the time we run `join()` the number of running threads would be
- * zero, even though no work has been done yet.
- */
-void idle_job::awaken_and_run(ThreadJob* job, size_t nthreads) {
-  xassert(job);
-  xassert(this_thread_index() == 0);
-  xassert(n_threads_running == 0);
-  xassert(prev_sleep_task->next_scheduler == nullptr);
-  xassert(curr_sleep_task->next_scheduler == nullptr);
-  int nth = static_cast<int>(nthreads) - 1;
-
-  std::swap(curr_sleep_task, prev_sleep_task);
-  n_threads_running += nth;
-  saved_exception = nullptr;
-
-  prev_sleep_task->next_scheduler = job;
-  prev_sleep_task->semaphore.signal(nth);
-  // enable_monitor(true);
-  master_worker->run_master(job);
-}
-
-
-// Wait until all threads go back to sleep (which would mean the job is done)
-void idle_job::join() {
-  xassert(this_thread_index() == 0);
-  // Busy-wait until all threads finish running
-  while (n_threads_running.load() != 0);
-
-  // Clear `.next_scheduler` flag of the previous sleep task, indicating that
-  // we no longer run in a parallel region (see `is_running()`).
-  prev_sleep_task->next_scheduler = nullptr;
-  // enable_monitor(false);
-
-  if (saved_exception) {
-    progress::manager->reset_interrupt_status();
-    std::rethrow_exception(saved_exception);
-  }
-
-  progress::manager->handle_interrupt();
-}
-
-
-void idle_job::set_master_worker(ThreadWorker* worker) noexcept {
-  master_worker = worker;
-}
-
-
-void idle_job::on_before_thread_removed() {
-  xassert(n_threads_running > 0);
-  n_threads_running--;
-}
-
-void idle_job::on_before_thread_added() {
-  n_threads_running++;
-}
-
-
-// Multiple threads may throw exceptions simultaneously, thus
-// need to protect access to `saved_exception` with a mutex.
-// In addition, `job->abort_execution()` is also protected,
-// just in case, ensuring that only one thread can call that
-// method at a time.
-//
-void idle_job::catch_exception() noexcept {
-  try {
-    std::lock_guard<std::mutex> lock(mutex);
-    if (!saved_exception) {
-      saved_exception = std::current_exception();
-    }
-    ThreadJob* current_job = prev_sleep_task->next_scheduler;
-    if (current_job) {
-      current_job->abort_execution();
-    }
-  } catch (...) {}
-}
-
-
-bool idle_job::is_running() const noexcept {
-  return (prev_sleep_task->next_scheduler != nullptr);
 }
